% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{all_classes}
\alias{all_classes}
\title{Combine all classes for an object in one character value}
\usage{
all_classes(x)
}
\arguments{
\item{x}{An object whose classes you want to examine.}
}
\value{
A character value listing the object's classes.
}
\description{
Converts the vector of classes for an object to a single comma
  delimited character value.
}
\details{
This function uses class(x) to extract the vector of classes for an
  object then collapses it into a single comma delimited character value.
}
\examples{
library(tibble)
x <- tibble(x = 1:3, y = letters[1:3])
class(x)
all_classes(x)

}
