% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Functions.R
\name{ci.rp}
\alias{ci.rp}
\title{Compute confidence intervals for Pearson correlations.}
\usage{
ci.rp(r, se, n, conf.level = 0.95, rn = NULL)
}
\arguments{
\item{r}{A numeric value for the correlation coefficient.}

\item{se}{A numeric value for the standard error of the correlation
coefficient.}

\item{n}{A numeric value for the sample size.}

\item{conf.level}{A numeric value for the confidence level of the returned
confidence interval, restricted to values between 0 and 1. Defaults to
0.95.}

\item{rn}{An optional character value to be used as a row name in the data
frame returned by this function.}
}
\value{
A data frame containing the correlation coefficient, standard error,
  confidence interval limits, z-statistic, p-value, s-value, BFB, and
  posterior probability of H1.
}
\description{
This function computes confidence intervals for Pearson
  correlations using the Fisher's z-transformation.
}
\details{
=============================================================================

This function computes the upper and lower bounds of an approximate,
  two-sided confidence interval based on the z-distribution, with the p-value
  also based on the z-test statistic. Additional statistics are also
  computed, namely the p-value, s-value, BFB, and posterior probability of
  H1.
}
\examples{
library(mvtnorm)
library(polycor)
set.seed(45284)
# Create a population correlation matrix.
R <- matrix(0, 4, 4)
R[upper.tri(R)] <- c(.2, .3, .4, .5, .6, .7)
diag(R) <- 1
R <- cov2cor(t(R) \%*\% R)
# Show population correlations.
round(R, 4)
# Simulate data with normal distributions and correlation structure R.
mydf <- rmvnorm(1000, mean = rep(0, 4), sigma = R)
mydf <- data.frame(mydf)
names(mydf) <- c("x1", "x2", "y1", "y2")
# Show sample correlations.
Rhat <- round(cor(mydf), 4)
Rhat
# Convert y1 & y2  into ordinal categorical variables.
mydf$y1 <- cut(mydf$y1, c(-Inf, .75, Inf))
mydf$y2 <- cut(mydf$y2, c(-Inf, -1, .5, 1.5, Inf))
# Pearson, polychoric, and polyserial correlations, ML estimates.
HC <- hetcor(mydf, ML = TRUE)
HC

# Pearson correlation, x1 & y1
ci.rp(r = HC$correlations[2,1], se = HC$std.errors[2,1],
      n = HC$n, rn = "x1 & x2")

}
\seealso{
\code{\link{p2s}} for s-values, \code{\link{p2bfb}} for BFBs, and
  \code{\link{p2pp}} for posterior probabilities.
}
