% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Functions.R
\name{ci.rpc}
\alias{ci.rpc}
\title{Compute confidence intervals for polychoric and polyserial
  correlations.}
\usage{
ci.rpc(r, se, rn = NULL)
}
\arguments{
\item{r}{A numeric value for the correlation coefficient.}

\item{se}{A numeric value for the standard error of the correlation
coefficient.}

\item{rn}{An optional character value to be used as a row name in the data
frame returned by this function.}
}
\value{
A data frame containing the correlation coefficient, standard error,
  confidence interval limits, z-statistic, p-value, s-value, BFB, and
  posterior probability of H1.
}
\description{
This function compute confidence intervals for polychoric and
  polyserial correlations using the Fisher's z-transformation.
}
\details{
This function implements the computations suggested by Raykov and
  Marcoulides (2011, p. 112) to obtain the upper and lower bounds of the
  confidence interval. Additional statistics are also computed, namely the
  s-value, BFB, and posterior probability of H1.
}
\references{
Raykov, T., & Marcoulides, G. A. (2011). Introduction to
  psychometric theory. New York, NY: Routledge.
}
\seealso{
\code{\link{p2s}} for s-values, \code{\link{p2bfb}} for BFBs, and
  \code{\link{p2pp}} for posterior probabilities.
}
