% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{CookDco}
\alias{CookDco}
\title{Compute a cutoff value for Cook's D values}
\usage{
CookDco(x)
}
\arguments{
\item{x}{A glm() model fit object.}
}
\value{
A numeric value for the cutoff.
}
\description{
This function computes a cutoff for Cook's D values from
  a glm() model fit.
}
\details{
=============================================================================

Fox (1997, p. 281) suggested a cutoff value for identifying
  observations with high Cook's D values in a GLM model. This function
  computes that cutoff. Examining observations with Cook's D values larger
  than the cutoff may be warranted.
}
\examples{
 m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
 CookDco(m1)

}
\references{
Fox, J. (1997). Applied regression analysis, linear models, and
  related methods. Thousand Oaks, CA: Sage Publications.
}
