% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_num.R
\name{display_num}
\alias{display_num}
\title{Display Numbers}
\usage{
display_num(
  x,
  upper = 1e+06,
  lower = 0.001,
  sci.format = "\%.2e",
  default = "\%.3f"
)
}
\arguments{
\item{x}{A numeric value (or vector).}

\item{upper}{A numeric value at or above which the number should be displayed
in scientific notation.}

\item{lower}{A numeric value below which the number should be displayed
in scientific notation.}

\item{sci.format}{A scientific notation format value suitable for use by
sprintf().}

\item{default}{A numeric format value suitable for use by sprintf().}
}
\value{
A character string for the formatted number.
}
\description{
This function converts very large and very small numeric values
  into a convenient format by using scientific notation
}
\details{
This function formats the number via sprintf(), using the
  sci.format when the number is larger than the upper parameter or smaller
  than the lower parameter. Otherwise it uses the default format. My main use
  for it is to display small p-values in a compact format without resorting
  to  imprecise categorizations like "< 0.001".
}
\examples{
display_num(c(2150000, 155, 0.025, 0.00006))

}
