% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{extract_glm}
\alias{extract_glm}
\title{Extract a custom texreg object from a glm() model}
\usage{
extract_glm(model, digits = 2)
}
\arguments{
\item{model}{A logistic regression model fit via glm(family = binomial).}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the result. Defaults to NULL, which does not round the result.}
}
\value{
A texreg object for the model.
}
\description{
Extracts a custom texreg object from a glm() model via
  broom::tidy() and broom::glance(), adding more goodness of fit statistics
  along the way.
}
\details{
=============================================================================

This function facilitate using [texreg::texreg()] to create tables
  summarizing logistic regression models that contain all the default
  goodness of fit statistics supplied by [broom::glance()], plus additional
  fit statistics generated by the piercer package functions [brier()],
  [pseudoR2()], and [R2Dev()]. The code was adapted from an online answer
  posted at https://stackoverflow.com/a/52037395.
}
\examples{
m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
extract_glm(m1)
extract_glm(m1, digits = 3)

}
\references{
Cameron, A. C., & Windmeijer, F. A. G. (1997). An R-squared
  measure of goodness of fit for some common nonlinear regression models.
  Journal of Econometrics, 77(2), 329-342. doi:10.1016/S0304-4076(96)01818-0

  Fox, J. (1997). Applied regression analysis, linear models, and related
  methods. Thousand Oaks, CA: Sage Publications.

  Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013). Applied
  logistic regression (3rd ed.). Hoboken, NJ: John Wiley & Sons, Inc.
}
\seealso{
See [brier()] for Brier scores, [pseudoR2()] for pseudo-R^2,
  [R2Dev()] for R^2 based on deviance. See [broom::tidy()] and
  [broom::glance()] for details about using the broom package to extract tidy
  summaries of models, and [texreg::texreg()] for details about using the
  texreg package to build tables summarizing regression models.
}
