% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_details.R
\name{file_details}
\alias{file_details}
\title{File Details}
\usage{
file_details(x)
}
\arguments{
\item{x}{A character value (or vector) of file paths.}
}
\value{
A tibble with one row per file path in x containing the file name,
  size, and date-time last modified.
}
\description{
This function shows details about one or more files in a tibble.
}
\details{
This function extracts the file name, size, and date-time last
  modified into a tibble.
}
