% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Functions.R
\name{Fisher_z2r}
\alias{Fisher_z2r}
\title{Fisher's z to r transformation.}
\usage{
Fisher_z2r(z)
}
\arguments{
\item{z}{A numeric value (or vector) for the z-score(s).}
}
\value{
A numeric value (or vector) for the correlation coefficient(s).
}
\description{
This function uses Fisher's z to r transformation to convert a
  z-score into a correlation.
}
\details{
=============================================================================

This function uses Fisher's z to r transformation to convert a
  correlation into a z-score. Missing values of z result in NA values.
  This is also called the hyperbolic tangent transformation (tanh), so this
  is a wrapper for the base::tanh() function.
}
\examples{
Fisher_z2r(.3)
Fisher_z2r(c(-Inf, -2.64, NA, 0, 0.55, 2.64, Inf))

}
\references{
To be added later.
}
\seealso{
\code{\link{base::tanh}} for the base function used in this
  function. \code{\link{Fisher_r2z}} for the inverse transformation that
  converts r back to z.
}
