% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE_Functions.R
\name{geep}
\alias{geep}
\title{Generalized estimating equation power for testing a main effect of
  treatment given a specific sample size (N, number of clusters)}
\usage{
geep(p0, p1, r, rho, n_e = 1, n_s = 1, pi_c = 0.5, alpha = 0.05, N)
}
\arguments{
\item{p0}{A numeric value for \eqn{p_0}{p_0}: the probability of the outcome
in the control arm.}

\item{p1}{A numeric value for \eqn{p_1}{p_1}: the probability of the outcome
in the treatment arm.}

\item{r}{A numeric value for \eqn{r}{r}: the correlation between evaluations
from the same subject.}

\item{rho}{A numeric value for \eqn{\rho}{rho}: the correlation between
outcome evaluations from different subjects in the same cluster.}

\item{n_e}{A numeric vector for \eqn{n_e}{n_e}: the number of outcome
evaluations per subject (level 1 observations of the binary outcome
variable), which is assumed to be constant across subjects.}

\item{n_s}{A numeric vector for \eqn{n_s}{n_s}: the number of subjects (level
2 units) per cluster (i.e., the cluster size), which is assumed to be
constant across clusters.}

\item{pi_c}{A numeric value for \eqn{\pi}{pi}: the proportion of clusters
in the control arm.}

\item{alpha}{A numeric value for \eqn{\alpha}{alpha}: the Type I error rate.}

\item{N}{A numeric vector for the total number of clusters in the sample.}
}
\value{
A data frame.
}
\description{
Compute the power for testing a main effect of a treatment in a
  2-arm parallel groups design via a 3-level generalized estimating equation
  with binary outcome data, given sample size (N, number of clusters).
}
\details{
=============================================================================

This function is useful for examining a range of scenarios. Using
  vectors with legnth > 1 for parameters such as n_s, n_e, and N will yield
  power estimates for multiple scenarios in the data frame returned.
}
\references{
Teerenstra, S., Lu, B., Preisser, J. S., van Achterberg, T., &
  Borm, G. F. (2010). Sample size considerations for GEE analyses of
  three-level cluster randomized trials. Biometrics, 66(4), 1230-1237.
  doi:10.1111/j.1541-0420.2009.01374.x
}
