% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{git_report}
\alias{git_report}
\title{Report on Git Repository Status}
\usage{
git_report(path = here())
}
\arguments{
\item{path}{A character value containing the path to a folder containing a
git repository, defaulting to the result of here::here().}
}
\value{
A data frame.
}
\description{
This function produces a short report about whether the current
  file is located in a Git repository and, if so, what commit it is at, and
  the status of the repository.
}
\details{
This function facilitates concisely adding version control output
  to report scripts. It is basically a wrapper that calls git2r::repository()
  and git2r::status() after verifying that you are in a Git repository.
}
\examples{
git_report()

}
