% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{hatco}
\alias{hatco}
\title{Compute a cutoff value for leverage hat values}
\usage{
hatco(x)
}
\arguments{
\item{x}{A glm() model fit object.}
}
\value{
A numeric value for the cutoff.
}
\description{
This function computes a cutoff for leverage hat values from
  a glm() model fit.
}
\details{
=============================================================================

Fox (1997, p. 280) suggested a cutoff value for identifying
  observations with high leverage hat values in a GLM model. This function
  computes that cutoff. Examining observations with hat values larger than
  the cutoff may be warranted.
}
\examples{
 m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
 hatco(m1)

}
\references{
Fox, J. (1997). Applied regression analysis, linear models, and
  related methods. Thousand Oaks, CA: Sage Publications.
}
