% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{InfCases}
\alias{InfCases}
\title{Identify influential cases (high leverage and high Cook's D)}
\usage{
InfCases(x, digits = 3)
}
\arguments{
\item{x}{A glm() model fit object.}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the result. Defaults to 3.}
}
\value{
A data frame showing observations that have both high leverage and
  high Cook's D.
}
\description{
This function identifies influential cases from a glm() model
  by finding those that exceed cutoffs for high leverage and high Cook's D.
}
\details{
=============================================================================

Fox (1997, p. 280-281) suggested cutoff values for identifying
  observations with high leverage hat values and high Cook's D values in a
  GLM model. Listing these influential observations is a quick way to inspect
  a model fit.
}
\examples{
 m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
 InfCases(m1)

}
\references{
Fox, J. (1997). Applied regression analysis, linear models, and
  related methods. Thousand Oaks, CA: Sage Publications.
}
\seealso{
\code{\link{hatco}} for leverage cutoff, \code{\link{CookDco}} for
  Cook's D cutoff.
}
