% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
A vector of estimated probabilities.
}
\description{
This function performs the inverse logit transformation, which
  converts continuous values to the range (0, 1).
}
\details{
=============================================================================

The inverse logit transform is useful when you want to convert
  an estimate from the log odds (logit) scale back into a probability. That
  may happen when working with logistic regression models.
}
\examples{
invlogit(0)
round(invlogit(-7:7), 3)

}
