% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{lrcm}
\alias{lrcm}
\title{Logistic regression classification measures}
\usage{
lrcm(
  roc,
  x = "best",
  best.method = "youden",
  transpose = FALSE,
  ret = "all",
  seed,
  ...
)
}
\arguments{
\item{roc}{a \dQuote{roc} object from the
	\code{\link[pROC]{roc}} function, or a \dQuote{smooth.roc} object from the
	\code{\link[=smooth.roc]{smooth}} function.
  }

\item{x}{
		the coordinates to look for. Numeric (if so, their meaning is
		defined by the \code{input} argument) or one of \dQuote{all} (all
		the points of the ROC curve), \dQuote{local maximas} (the local
		maximas of the ROC curve) or \dQuote{best} (see \code{best.method}
		argument). If missing or \code{NULL}, defaults to \dQuote{all}.
  }

\item{best.method}{if \code{x="best"}, the method to determine the
    best threshold. See details in the \sQuote{Best thresholds} section.
  }

\item{transpose}{when returning a matrix (\code{as.list=FALSE}), whether
  	to return of thresholds in columns (\code{TRUE}) or rows (\code{FALSE}).
  	The current default is \code{TRUE} and has been so since the initial
  	version. However an upcoming version will change the default to \code{FALSE},
  	and a warning is printed if the argument is not passed explicitly.
  	See \link[pROC]{coords_transpose} for more details about upcoming changes.
  }

\item{ret}{The coordinates to return. See "Valid \code{ret} arguments"
  	section below. Alternatively, the single value \dQuote{all} can be used to return
    every coordinate available.
  }

\item{seed}{A single number to be passed to set.seed(), which is used to
ensure reproducibility of the bootstrapped confidence intervals.}

\item{...}{further arguments passed from other methods. Ignored.}
}
\value{
A data frame showing estimates and bootstrapped confidence intervals
  for various classification measures.
}
\description{
This function computes a variety of classification measures
  relevant to logistic regression models.
}
\details{
=============================================================================

This function is a convnient wrapper around pROC::coords() and
  pROC::ci.coords(). Most of the arguments are passed along to those
  functions. What lrcm() does is just gather the results of both estimates
  and bootstrapped confidence intervals into a data frame. See the details
  sections of documentetion for pROC::coords() and pROC::ci.coords() for more
  information. Some information mentioned in the descriptions of individual
  parameters listed above is absent here because these parameters are
  inherited from pROC. It is more efficient to refer you to the pROC
  documentation than to retype it.
}
\examples{
 library(pROC)
 m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
 set.seed(4921) # For reproducibility of bootstrap estimates.
 rocm1 <- roc(m1$y ~ predict(m1, type = "response"), ci = TRUE,
              direction = "<", ci.method = "bootstrap")
 print(rocm1)
 lrcm(rocm1, seed = 563)

}
\references{
Robin, X., Turck, N., Hainard, A., Tiberti, N., Lisacek, F.,
  Sanchez, J.-C., & Müller, M. (2011). pROC: an open-source package for R and
  S+ to analyze and compare ROC curves. BMC Bioinformatics, 12, 77.
  doi:10.1186/1471-2105-12-77

  Youden, W. J. (1950). Index for rating diagnostic tests. Cancer, 3(1),
  32-35.
  doi:10.1002/1097-0142(1950)3%3A1%3C32%3A%3AAID-CNCR2820030106%3E3.0.CO%3B2-3
}
\seealso{
\code{\link[pROC]{roc}}, \code{\link[pROC]{coords}},
  \code{\link[pROC]{ci.coords}}, \code{\link[base]{set.seed}}.
}
