% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{move_file}
\alias{move_file}
\title{Move files}
\usage{
move_file(from, to, ...)
}
\arguments{
\item{from}{Character vector, containing source file names or paths to be
moved.}

\item{to}{Character vector, containing destination file names or paths.}

\item{...}{Arguments passed to `file.copy`}
}
\description{
Moves files to another location.
}
\details{
This function is a wrapper for both `file.copy` and `file.remove`
  that first copies the source file(s) to the destination, then removes them
  from their original location only if all source files were successfully
  copied.
}
