% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P-value_Functions.R
\name{p2bfb}
\alias{p2bfb}
\title{Convert p-values into Bayes Factor Bounds (BFBs)}
\usage{
p2bfb(p, digits = NULL)
}
\arguments{
\item{p}{A numeric p-value (or numeric vector of p-values), obtained from
running a statistical model or test. All values should be proportions
within the closed unit interval (0 <= p <= 1).}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the BFB. Defaults to NULL, which does not round the result.}
}
\value{
A numeric vector of BFB values of the same length as p.
}
\description{
This function converts p-values into Bayes Factor Bounds (BFBs),
  which are the data-based odds of the alternative hypotheis H1 being true to
  the null hypothesis H0 being true and the upper bound on the Bayes Factor.
}
\details{
=============================================================================

Bayes Factor Bounds (BFBs)  are the data-based odds of the
  alternative hypotheis H1 being true to the null hypothesis H0 being true
  (Benjamin & Berger, 2019). They are the upper bound on the Bayes Factor
  for a given test. The BFBs are most useful when the p-value lies in the
  open unit interval (0 < p < 1). Plugging in p = 0 or p = 1 will return
  NaN or -Inf, respectively.
}
\examples{
p2bfb(.05)
p2bfb(c(.1, .05, .01, .005, .001, .0001, .00001), digits = 2)
p2bfb(0)

}
\references{
Benjamin, D. J., & Berger, J. O. (2019). Three recommendations
  for improving the use of p-values. The American Statistician, 73
  (Supplement 1), 186-191. doi:10.1080/00031305.2018.1543135

  Wasserstein, R. L., Schirm, A. L., & Lazar, N. A. (2019). Moving to a world
  beyond "p < .05". The American Statistician, 73(Supplement 1), 1-19.
  doi:10.1080/00031305.2019.1583913
}
