% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Functions.R
\name{p2odds}
\alias{p2odds}
\title{Convert from probability to odds.}
\usage{
p2odds(p)
}
\arguments{
\item{p}{A numeric value (or vector) for the probabilities to be converted.
NA values are acceptable, but will return NA. All numeric values must fall
in the closed unit interval, [0, 1].}
}
\value{
A numeric value (or vector) for the odds values.
}
\description{
This function converts probability values into odds values.
}
\details{
=============================================================================

This function uses the formula odds = p/(1 - p).
}
\examples{
p2odds(.3)
p2odds(c(0, .1, .5, .75, .99, 1))

}
\references{
To be added later.
}
