% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform_Functions.R
\name{p2or}
\alias{p2or}
\title{Convert from a pair of probabilities to an odds ratio.}
\usage{
p2or(p0, p1)
}
\arguments{
\item{p0}{A numeric value (or vector) for the probabilities of success in
the reference group or category (i.e., the denominator for the odds ratio).
NA values are acceptable, but will return NA. All numeric values must fall
in the closed unit interval, [0, 1].}

\item{p1}{A numeric value (or vector) for the probabilities of success in
the focal group or category (i.e., the numerator for the odds ratio).
NA values are acceptable, but will return NA. All numeric values must fall
in the closed unit interval, [0, 1].}
}
\value{
A numeric value (or vector) for the odds ratio values.
}
\description{
This function converts a pair of probability values into an odds
  ratio.
}
\details{
=============================================================================

This function uses the formula OR = (p1/(1 - p1))/(p0/(1 - p0)).
}
\examples{
p2or(p0 = .3, p1 = .6)
p2or(p0 = c(.3, .4), p1 = c(.6, .5))

}
\references{
To be added later.
}
