% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piercer-package.R
\docType{package}
\name{piercer-package}
\alias{piercer}
\alias{piercer-package}
\title{piercer: Functions for Research and Statistical Computing}
\description{
Personal package of miscellaneous functions for use in Steven J. 
    Pierce's research and statistical collaboration and consulting work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjpierce/piercer}
}

}
\author{
\strong{Maintainer}: Steven J. Pierce \email{pierces1@msu.edu} (\href{https://orcid.org/0000-0002-0679-3019}{ORCID}) [copyright holder]

}
\keyword{internal}
