% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_functions.R
\name{pseudoR2}
\alias{pseudoR2}
\title{Compute pseudo-\eqn{R^2} for a logistic regression model}
\usage{
pseudoR2(x, digits = NULL)
}
\arguments{
\item{x}{A logistic regression model fit via glm(family = binomial).}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the result. Defaults to NULL, which does not round the result.}
}
\value{
A numeric value for the pseudo-\eqn{R^2}.
}
\description{
Computes a simple pseudo-\eqn{R^2} measure for a logistic
  regression model.
}
\details{
=============================================================================

This pseudo-R2 measure is just the Pearson correlation between the
  observed and fitted values from the logistic regression model, as discussed
  by Hosmer, Lemeshow, & Sturdivant (2013, p. 182).
}
\examples{
m1 <- glm(formula = vs ~ wt + disp, family = binomial, data = mtcars)
pseudoR2(m1)
pseudoR2(m1, digits = 2)

}
\references{
Hosmer, D. W., Lemeshow, S., & Sturdivant, R. X. (2013). Applied
  logistic regression (3rd ed.). Hoboken, NJ: John Wiley & Sons, Inc.
}
