% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_Functions.R
\name{r.ps}
\alias{r.ps}
\title{Compute confidence intervals for a set of polyserial correlations.}
\usage{
r.ps(x, cont, ord, digits = NULL, pdigits = NULL)
}
\arguments{
\item{x}{A hetcor object produced by hetcor().}

\item{cont}{A character vector of names for continuous variables.}

\item{ord}{A character vector of names for ordinal variables.}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the correlation, SE, CI bounds, z-statistic, s-value, BFB, and
posterior probability. Defaults to NULL, which does not round the result.}

\item{pdigits}{An integer specifying the number of decimal places to used when
rounding the p-value. Defaults to NULL, which does not round the result.}
}
\value{
A data frame containing the results.
}
\description{
This function computes confidence intervals for polyserial
  correlations obtained from a hetcor object.
}
\details{
=============================================================================

This function applies ci.pc() to all the polyserial correlations
  in the hetcor object supplied by the user.
}
\references{
Raykov, T., & Marcoulides, G. A. (2011). Introduction to
  psychometric theory. New York, NY: Routledge.
}
\seealso{
\code{\link{ci.rpc}} for the function used to get the CIs,
  \code{\link{p2s}} for s-values, \code{\link{p2bfb}} for BFBs, and
  \code{\link{p2pp}} for posterior probabilities.
}
