% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NL-SEM_Reliability.R
\name{rxx.NL}
\alias{rxx.NL}
\title{Non-linear SEM composite reliability coefficient}
\usage{
rxx.NL(lambda, tau, fcor = matrix(1, ncol = 1), digits = NULL)
}
\arguments{
\item{lambda}{A factor loading matrix (items x factors). See details below.}

\item{tau}{A threshold matrix (items x thresholds, same set for all items).}

\item{fcor}{A latent factor correlation matrix (factors x factors), which
defaults to a 1 x 1 matrix: [cor = 1]}

\item{digits}{An integer specifying the number of decimal places to used when
rounding the result. Defaults to NULL, which does not round the result.}
}
\value{
A numeric value for the reliability.
}
\description{
Computes a reliability estimate for a latent factor measured by
  ordinal indicator items from confirmatory factor analysis (CFA) parameter
  estimates.
}
\details{
=============================================================================

This function computes Green & Yang's (2009) nonlinear SEM
  reliability coefficient from CFA with ordinal indicators (Equation 21).
  The code was adapted from Yang & Green's (2015) SAS macro, found online at
  http://myweb.fsu.edu/yyang3/files/examplecodes.txt. Note that Green &
  Yang's (2009) Equation 21 matches Yang & Green's (2015) Equation 9.

  This code was written assuming one would use Mplus to estimate the CFA, so
  some terminology here reflects that. If you are using WLSMV estimation with
  latent variance fixed to 1 and all indicator loadings freely estimated,
  then unstandardized & standardized estimates are identical (you can use
  either one as the lambda input to this function). However, if you are using
  Bayesian estimation with PX parameterization, use the STDYX estimates.
}
\references{
Green, S. B., & Yang, Y. (2009). Reliability of summed item
  scores using structural equation modeling: An alternative to coefficient
  alpha. Psychometrika, 74(1), 155-167. doi:10.1007/s11336-008-9099-3

  Yang, Y., & Green, S. B. (2015). Evaluation of structural equation modeling
  estimates of reliability for scales with ordered categorical items.
  Methodology: European Journal of Research Methods for the Behavioral and
  Social Sciences, 11(1), 23-34. doi:10.1027/1614-2241/a000087
}
