% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimTools.R
\name{sim_2arm_prepost}
\alias{sim_2arm_prepost}
\title{Simulate Pretest-Postest data}
\usage{
sim_2arm_prepost(
  Npg = 5,
  VNames = c("x", "y"),
  R = matrix(data = c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE),
  MV.C = c(x = 0, y = 0),
  MV.T = c(x = 0, y = 1),
  SV.C = c(x = 1, y = 1),
  SV.T = c(x = 1, y = 1)
)
}
\arguments{
\item{Npg}{An integer value for the number of subjects per group.}

\item{VNames}{A character vector with the names of the variables, in the
order they will be used in the population matrix row and column names. These
will be used as dimnames. It defaults to the list of two variable names (x
and y), where x is the pretest data and y is the posttest data.}

\item{R}{A population correlation matrix. This will be used by
mvtnorm::rmvnorm(). The default value is arbitrary.}

\item{MV.C}{A numeric vector of means for the control group. This must be the
same length as VNames and in the same order. This will be used by
mvtnorm::rmvnorm(). It defaults to means on 1.}

\item{MV.T}{A numeric vector of means for the treatment group. This must be
the same length as VNames and in the same order. This will be used by
mvtnorm::rmvnorm().}

\item{SV.C}{A numeric vector of standard deviations for the control group.
This must be the same length as VNames and in the same order. This will be
used by mvtnorm::rmvnorm().}

\item{SV.T}{A numeric vector of standard deviations for the control group.
This must be the same length as VNames and in the same order. This will be
used by mvtnorm::rmvnorm().}
}
\value{
A tibble or a data frame.
}
\description{
Generates simulated data for use in study sample size planning
  for a 2 arm, pretest/posttest design.
}
\details{
=============================================================================

This function generates simulated data for use in study sample size
  planning, based on the supplied inputs. It is strictly designed to work
  with a two-group, pretest/posttest design where the planned analysis would
  be a a simple ANCOVA model. The simulation creates separate data frames for
  the control and treatment groups that each contain pretest data (x) and
  posttest data (y). The two groups always have equal sample sizes.

  PID is a person level identifier, and Arm indicates whether the person
  belongs to the control or treatment group.
}
\examples{
# Load packages.
library(dplyr)
library(mvtnorm)
library(MBESS)
# Default settings.
set.seed(1528)
sim_2arm_prepost()

}
