% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_missingness.R
\name{var_missingness}
\alias{var_missingness}
\title{Variables Missingness}
\usage{
var_missingness(x)
}
\arguments{
\item{x}{A dataframe for which you want a summary of variable-level missingness.}
}
\value{
A data frame with one row per variable in the the input data frame
  x and 7 columns (position and name of the variable, total number of values,
  plus number and percent of valid values, number and percentage of missing
  values.
}
\description{
This function summarizes the numbers of valid (non-missing) and
  missing values and the percent of missing values for each variable in a
  data frame or tibble.
}
\details{
This function summary of how many valid and missing values there are
  for each variable in a data frame.
}
\examples{
var_missingness(cars)

}
