% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEE_Functions.R
\name{vif_gee_crt}
\alias{vif_gee_crt}
\title{Variance inflation factor for a generalized estimating equation (GEE)
  analysis of a 3-level cluster randomized trial}
\usage{
vif_gee_crt(r, rho, n_e = 1, n_s = 1)
}
\arguments{
\item{r}{A numeric value for \eqn{r}{r}: the correlation between evaluations
from the same subject.}

\item{rho}{A numeric value for \eqn{\rho}{rho}: the correlation between
outcome evaluations from different subjects in the same cluster.}

\item{n_e}{A numeric vector for \eqn{n_e}{n_e}: the number of outcome
evaluations per subject (level 1 observations of the binary outcome
variable), which is assumed to be constant across subjects.}

\item{n_s}{A numeric vector for \eqn{n_s}{n_s}: the number of subjects (level
2 units) per cluster (i.e., the cluster size), which is assumed to be
constant across clusters.}
}
\value{
A numeric value.
}
\description{
Compute the variance inflation factor applicable to a maximum
  likelihood analysis via a generalized estimating equation (GEE) for a
  three-level cluster randomized trial.
}
\details{
=============================================================================

This function is useful for power analysis calculations based on
  the methods described in Teerenstra et. al (2010).
}
\examples{

vif_gee_crt(r = .18, rho = .02, n_e = 15, n_s = 22.1)

}
\references{
Teerenstra, S., Lu, B., Preisser, J. S., van Achterberg, T., &
  Borm, G. F. (2010). Sample size considerations for GEE analyses of
  three-level cluster randomized trials. Biometrics, 66(4), 1230-1237.
  doi:10.1111/j.1541-0420.2009.01374.x
}
