% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility_Functions.R
\name{which_latex}
\alias{which_latex}
\title{Check which LaTeX software will compile PDF output}
\usage{
which_latex()
}
\value{
A character value describing which LaTeX software will be used.
}
\description{
Check which LaTeX software (TinyTeX vs. other LaTeX
  distribution) will be used to compile PDF output.
}
\details{
This function facilitates checking whether the computer running the
  script will use TinyTeX or some other LaTeX software to compile PDF files
  from LaTeX files. It assumes that if TinyTeX is installed, that will be
  used in preference to some alternative LaTeX software (which would be
  MiKTeX on CSTAT's virtual server).
}
\examples{
which_latex()

}
